Don't extend the Symfony class anymore, as calculating the Specificity is the task of the Processor
