+++ src/Css/Property/Processor.php
@@ -2,7 +2,7 @@
 
 namespace TijsVerkoyen\CssToInlineStyles\Css\Property;
 
-use TijsVerkoyen\CssToInlineStyles\Css\Specificity\Specificity;
+use Symfony\Component\CssSelector\Node\Specificity;
 
 class Processor
 {
